+++ src/functions/AlwaysSpawn.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-using namespace geode::prelude;
-
-extern void alwaysSpawnFunc(PlayerObject* player);
-extern std::array<bool, 2> gate;
-extern std::array<int, 2> landSwitch;