+++ src/functions/Core.hpp
@@ -1,2 +1,0 @@
-#pragma once
-extern void core(PlayerObject* player);