+++ src/functions/util.hpp
@@ -1,3 +1,0 @@
-#pragma once
-extern bool check(PlayerObject* player, bool detectPlayer);
-extern int index(PlayerObject* player);