+++ src/globals/settings.cpp
@@ -1,0 +1,16 @@
+#include <Geode/Geode.hpp>
+#include <Geode/modify/PlayerObject.hpp>
+#include "settings.hpp"
+
+using namespace geode::prelude;
+
+bool g_isModEnabled = true;
+bool g_isLegacyTracking = true;
+bool g_isLegacyScaling = false;
+bool g_isTrueValues = true;
+bool g_isAlwaysSpawn = true;
+bool g_isNoRotation = true;
+bool g_affectP1 = true;
+bool g_affectP2 = true;
+bool g_centerDashX = true;
+bool g_centerDashY = true;