+++ src/globals/settings.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+extern bool g_isModEnabled;
+extern bool g_isLegacyTracking;
+extern bool g_isLegacyScaling;
+extern bool g_isTrueValues;
+extern bool g_isAlwaysSpawn;
+extern bool g_isNoRotation;
+extern bool g_centerDashX;
+extern bool g_centerDashY;
+extern bool g_affectP1;
+extern bool g_affectP2;