+++ src/util.hpp
@@ -1,0 +1,3 @@
+#pragma once
+extern bool isValid(PlayerObject* player, bool detectPlayer);
+extern int playerIndex(PlayerObject* player);